requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.11
djangorestframework==3.8.2
drf-yasg==1.6.1
flex==6.13.1
future==0.16.0
idna==2.6
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
MarkupSafe==1.0
openapi-codec==1.3.2
pytz==2018.4
PyYAML==3.12
requests==2.18.4
rfc3987==1.3.7
ruamel.yaml==0.15.37
six==1.11.0
strict-rfc3339==0.7
swagger-spec-validator==2.1.0
uritemplate==3.0.0
urllib3==1.22
validate-email==1.3
